refactor(impact): extract ProjectDiscovery subcomponents and prune dead barrel export
- Split FilterSelect, IndicatorSliders, DiscoveryResults, DiscoveryResultCard
  into sibling files so ProjectDiscovery.tsx drops from 636 to 308 lines,
  clearing the 600-line quality-gate size budget.
- Share the IndicatorDistribution type from services/projectDiscovery instead
  of redefining it locally, so the exported type is consumed (knip).
- Remove the now-unused getCommunityCategories re-export from the community
  barrel; the hook uses getCommunityCategoriesOrThrow and all other callers
  import getCommunityCategories directly from getCommunityData.
