+++ utilities/queries/v2/community.ts
@@ -1,6 +1,5 @@
 // Re-export from canonical location
 export {
-  getCommunityCategories,
   getCommunityCategoriesOrThrow,
   getCommunityDetails,
   getCommunityDetailsV2,
